+++ src/utils/regexps.mjs
@@ -4,3 +4,4 @@ export const attributeRegExp = /[\w-]+=['"]?$/
 export const booleanAttributeRegExp = /\?([\w-]+)=['"]?$/
 export const eventListenerRegExp = /@([\w-]+)=['"]?$/
 export const dataEventIdJSRegExp = /^eventId[\d]{1,}$/
+export const styleAttributeRegExp = /style=['"]?$/